High-performance 'Morlet'-wavellet on 'RAVE' subject

This function is intended to be called within builtin pipelines. Please check https://rave.wiki to see details.

wavelet_rave_subject(
  subject,
  electrode,
  frequencies,
  kernel_cycles,
  downsample_rate = 100,
  demean = TRUE,
  signal_source = "/notch/{block}"
)

Arguments

subject
character of subject ID or RAVESubject instance

electrode
length of one, which electrode should be applied

frequencies, kernel_cycles, demean
see morlet_wavelet

downsample_rate
down-sample rate after wavelet

signal_source
which 'HDF5' name should be used. Current default is "/notch/{block}" for 'Notch' filtered signals

Value

Current time-stamp. The wavelet power and phase will be saved to 'RAVE' structure automatically