Compute quantiles

Compute quantiles

Usage

fast_quantile(x, prob = 0.5, na.rm = FALSE, ...)

fast_median(x, na.rm = FALSE, ...)

fast_mvquantile(x, prob = 0.5, na.rm = FALSE, ...)

fast_mvmedian(x, na.rm = FALSE, ...)

Arguments

x

numerical-value vector for fast_quantile and fast_median, and column-major matrix for fast_mvquantile and fast_mvmedian

prob

a probability with value from 0 to 1

na.rm

logical; if true, any NA are removed from x before the quantiles are computed

...

reserved for future use

Value

fast_quantile and fast_median calculate univariate quantiles (single-value return); fast_mvquantile and fast_mvmedian

calculate multivariate quantiles (for each column, result lengths equal to the number of columns).

Examples

fast_quantile(runif(1000), 0.1)
#> [1] 0.08398884
fast_median(1:100)
#> [1] 50.5

x <- matrix(rnorm(100), ncol = 2)
fast_mvquantile(x, 0.2)
#> [1] -0.4647119 -0.9357731
fast_mvmedian(x)
#> [1] 0.25158133 0.02348051

# Compare speed for vectors (usually 30% faster)
x <- rnorm(10000)
microbenchmark::microbenchmark(
  fast_median = fast_median(x),
  base_median = median(x),
  # bioc_median = Biobase::rowMedians(matrix(x, nrow = 1)),
  times = 100, unit = "milliseconds"
)
#> Unit: milliseconds
#>         expr    min      lq     mean median     uq    max neval
#>  fast_median 0.0903 0.13830 0.166139 0.1584 0.1966 0.2671   100
#>  base_median 0.2279 0.23595 0.278500 0.3023 0.3133 0.3951   100

# Multivariate cases
# (5~7x faster than base R)
# (3~5x faster than Biobase rowMedians)
x <- matrix(rnorm(100000), ncol = 20)
microbenchmark::microbenchmark(
  fast_median = fast_mvmedian(x),
  base_median = apply(x, 2, median),
  # bioc_median = Biobase::rowMedians(t(x)),
  times = 10, unit = "milliseconds"
)
#> Unit: milliseconds
#>         expr    min     lq    mean median     uq    max neval
#>  fast_median 0.8454 0.8987 1.06123 0.9131 1.0291 1.7035    10
#>  base_median 3.4338 3.5670 3.97694 3.6541 4.1009 5.3023    10